octofit-tracker/backend/leaderboard/serializers.py
from rest_framework import serializers
from .models import LeaderboardEntry

class LeaderboardEntrySerializer(serializers.ModelSerializer):
    username = serializers.CharField(source='user.username', read_only=True)
    
    class Meta:
        model = LeaderboardEntry
        fields = ['id', 'user', 'username', 'period', 'rank', 'points', 
                  'period_start', 'period_end', 'created_at']
        read_only_fields = ['id', 'created_at']
