octofit-tracker/backend/requirements.txt
Django==4.2.26
djangorestframework==3.14.0
django-allauth==0.51.0
django-cors-headers==4.5.0
dj-rest-auth==2.2.6
